Remove application.ex and config
