--- starter_templates/redis/elixir/app/lib/application.ex
@@ -1,14 +1,0 @@
-defmodule Server.Application do
-  use Application
-
-  def start(_type, _args) do
-    children = [
-      # Uncomment this to pass the first stage
-      #
-      Server.accept()
-    ]
-
-    opts = [strategy: :simple_one_for_one, name: Server.Supervisor]
-    Supervisor.start_link(children, opts)
-  end
-end
