+++ starter_templates/redis/elixir/app/mix.exs
@@ -14,7 +14,7 @@ defmodule App.MixProject do
   def application do
     [
       extra_applications: [:logger],
-      mod: {Server.Application, []}
+      mod: {Server, []}
     ]
   end
 end
